+++ modules/i3s/src/i3s-content-loader.js
@@ -14,8 +14,7 @@ export const I3SContentLoader = {
   name: 'I3S Content (Indexed Scene Layers)',
   id: 'i3s-content',
   module: 'i3s',
-  // Return "true" after featureIds replaced with segmentationData in I3S-picking-app
-  worker: false,
+  worker: true,
   version: VERSION,
   mimeTypes: ['application/octet-stream'],
   parse,
@@ -27,9 +26,6 @@ export const I3SContentLoader = {
 
 async function parse(data, options) {
   const {tile, tileset} = options.i3s;
-  tile.content = tile.content || {};
-  tile.segmentationData = tile.segmentationData || null;
-  tile.userData = tile.userData || {};
   await parseI3STileContent(data, tile, tileset, options);
   return tile.content;
 }