+++ modules/i3s/test/helpers/i3s-nodepages-tiles.spec.js
@@ -26,17 +26,17 @@ test('I3SNodePagesTiles#Tile with content', async t => {
   t.ok(node1);
   t.equal(
     node1.contentUrl,
-    '@loaders.gl/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/1/geometries/0'
+    'https://raw.githubusercontent.com/visgl/loaders.gl/master/modules/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/1/geometries/0'
   );
   if (isBrowser) {
     t.equal(
       node1.textureUrl,
-      '@loaders.gl/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/1/textures/0_0_1'
+      'https://raw.githubusercontent.com/visgl/loaders.gl/master/modules/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/1/textures/0_0_1'
     );
   } else {
     t.equal(
       node1.textureUrl,
-      '@loaders.gl/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/1/textures/0'
+      'https://raw.githubusercontent.com/visgl/loaders.gl/master/modules/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/1/textures/0'
     );
   }
 
@@ -77,7 +77,8 @@ test('I3SNodePagesTiles#Layer without textures', async t => {
   t.end();
 });
 
-test('I3SNodePagesTiles#Tile should have mbs converted from obb', async t => {
+// Logic moved to parse-i3s.js to avoid calling extra conversion the center from cartographic to cartesian
+test.skip('I3SNodePagesTiles#Tile should have mbs converted from obb', async t => {
   const i3SNodePagesTiles = new I3SNodePagesTiles(TILESET_STUB(), {});
   const node1 = await i3SNodePagesTiles.formTileFromNodePages(1);
   t.ok(node1);
@@ -120,7 +121,7 @@ test('I3SNodePagesTiles#Select "dds" texture if it is supported', async t => {
     if (supportedFormats.has('dxt')) {
       t.equal(
         node.textureUrl,
-        '@loaders.gl/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/2/textures/0_0_1'
+        'https://raw.githubusercontent.com/visgl/loaders.gl/master/modules/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/2/textures/0_0_1'
       );
       t.deepEqual(i3SNodePagesTiles.textureDefinitionsSelectedFormats, [
         {name: '0_0_1', format: 'dds'}
@@ -128,7 +129,7 @@ test('I3SNodePagesTiles#Select "dds" texture if it is supported', async t => {
     } else {
       t.equal(
         node.textureUrl,
-        '@loaders.gl/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/2/textures/0'
+        'https://raw.githubusercontent.com/visgl/loaders.gl/master/modules/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/2/textures/0'
       );
       t.deepEqual(i3SNodePagesTiles.textureDefinitionsSelectedFormats, [
         {name: '0', format: 'jpg'}
@@ -137,7 +138,7 @@ test('I3SNodePagesTiles#Select "dds" texture if it is supported', async t => {
   } else {
     t.equal(
       node.textureUrl,
-      '@loaders.gl/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/2/textures/0'
+      'https://raw.githubusercontent.com/visgl/loaders.gl/master/modules/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/2/textures/0'
     );
     t.deepEqual(i3SNodePagesTiles.textureDefinitionsSelectedFormats, [{name: '0', format: 'jpg'}]);
   }
@@ -175,7 +176,7 @@ test('I3SNodePagesTiles#Switch off compressed textures', async t => {
     if (supportedFormats.has('dxt')) {
       t.equal(
         node.textureUrl,
-        '@loaders.gl/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/2/textures/0'
+        'https://raw.githubusercontent.com/visgl/loaders.gl/master/modules/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/2/textures/0'
       );
       t.deepEqual(i3SNodePagesTiles.textureDefinitionsSelectedFormats, [
         {name: '0', format: 'jpg'}
@@ -192,7 +193,7 @@ test('I3SNodePagesTiles#Should load DRACO geometry', async t => {
   t.ok(node1);
   t.equal(
     node1.contentUrl,
-    '@loaders.gl/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/1/geometries/1'
+    'https://raw.githubusercontent.com/visgl/loaders.gl/master/modules/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/1/geometries/1'
   );
 
   const tilesetJson = TILESET_STUB();
@@ -205,7 +206,7 @@ test('I3SNodePagesTiles#Should load DRACO geometry', async t => {
   const node12 = await i3SNodePagesTiles2.formTileFromNodePages(1);
   t.equal(
     node12.contentUrl,
-    '@loaders.gl/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/1/geometries/0'
+    'https://raw.githubusercontent.com/visgl/loaders.gl/master/modules/i3s/test/data/SanFrancisco_3DObjects_1_7/SceneServer/layers/0/nodes/1/geometries/0'
   );
 
   t.end();