+++ modules/i3s/test/lib/parsers/parse-i3s-tile-content.spec.js
@@ -36,7 +36,8 @@ test('ParseI3sTileContent#should load "dds" texture if it is supported', async t
       useDracoGeometry: false
     }
   });
-  const texture = result.content.texture;
+  const texture =
+    result.content.material.pbrMetallicRoughness.baseColorTexture.texture.source.image;
   if (isBrowser) {
     const supportedFormats = getSupportedGPUTextureFormats();
     if (supportedFormats.has('dxt')) {
@@ -99,7 +100,17 @@ test('ParseI3sTileContent#should have segmentationData', async t => {
     }
   });
   t.ok(result);
-  t.ok(result.segmentationData);
-  t.equal(result.segmentationData.length, 25638);
+  t.ok(result.content.segmentationData);
+  t.equal(result.content.segmentationData.length, 25638);
+  t.end();
+});
+
+test('ParseI3sTileContent#should generate mbs from obb', async t => {
+  const i3sTilesetData = TILESET_STUB();
+  const i3SNodePagesTiles = new I3SNodePagesTiles(i3sTilesetData, {});
+  const tile = await i3SNodePagesTiles.formTileFromNodePages(1);
+  t.ok(tile.mbs);
+  t.equals(tile.mbs.length, 4);
+  t.deepEquals(tile.mbs.slice(0, 3), tile.obb.center);
   t.end();
 });