+++ modules/mvt/src/lib/binary-vector-tile/vector-tile-feature.d.ts
@@ -25,7 +25,7 @@ export default class VectorTileFeature {
   constructor(pbf: Protobuf, end, extent, keys, values, firstPassData);
 
   loadGeometry(); any;
-  classifyRings(geom: {data: [number], lines: [number]}): [[number]];
+  classifyRings(geom: {data: [number], lines: [number]}): {data: [number], areas: [[number]], lines: [[number]]};
   toBinaryCoordinates(options: {x: number, y: number, z: number} | (([number], VectorTileFeature) => void)): object;
 }