+++ modules/tile-converter/src/i3s-converter/i3s-converter.js
@@ -57,7 +57,7 @@ export default class I3SConverter {
       tilesCount: 0,
       tilesWithAddRefineCount: 0
     };
-    this.validateBoundingVolumes = false;
+    this.validate = false;
     this.boundingVolumeWarnings = null;
   }
 
@@ -72,11 +72,11 @@ export default class I3SConverter {
     egmFilePath,
     token,
     draco,
-    validateBoundingVolumes
+    validate
   }) {
     this.conversionStartTime = process.hrtime();
     this.options = {maxDepth, slpk, sevenZipExe, egmFilePath, draco, token, inputUrl};
-    this.validateBoundingVolumes = validateBoundingVolumes;
+    this.validate = validate;
 
     console.log('Loading egm file...'); // eslint-disable-line
     this.geoidHeightModel = await load(egmFilePath, PGMLoader);
@@ -86,15 +86,12 @@ export default class I3SConverter {
       this.nodePages.useWriteFunction(writeFileForSlpk);
     }
 
-    const options = await this._fetchPreloadOptions();
-    const sourceTilesetJson = await load(inputUrl, CesiumIonLoader, options);
-
-    /* TODO/ib - get rid of confusing options warnings, move into options sub-object */
-    // const tilesetJson = await load(inputUrl, CesiumIonLoader, {
-    //   'cesium-ion': preloadOptions
-    // });
+    const fetchOptions = await this._fetchPreloadOptions();
+    const sourceTilesetJson = await load(inputUrl, CesiumIonLoader, {
+      fetch: fetchOptions
+    });
     // console.log(tilesetJson); // eslint-disable-line
-    this.sourceTileset = new Tileset3D(sourceTilesetJson, options);
+    this.sourceTileset = new Tileset3D(sourceTilesetJson, fetchOptions);
 
     await this._createAndSaveTileset(outputPath, tilesetName);
     await this._finishConversion({slpk, outputPath, tilesetName});
@@ -411,7 +408,10 @@ export default class I3SConverter {
    * @return {Promise<object[]>}
    */
   async _createNode(parentTile, sourceTile, parentId, level) {
-    this._checkAddRefinementTypeForTile(sourceTile);
+    if (this.validate) {
+      this._checkAddRefinementTypeForTile(sourceTile);
+    }
+
     await this._updateTilesetOptions();
     await this.sourceTileset._loadTile(sourceTile);
     const coordinates = convertCommonToI3SCoordinate(sourceTile, this.geoidHeightModel);
@@ -456,7 +456,7 @@ export default class I3SConverter {
         await this._writeResources(resources, node.path);
       }
 
-      if (this.validateBoundingVolumes) {
+      if (this.validate) {
         this.boundingVolumeWarnings = validateNodeBoundingVolumes(node);
 
         if (this.boundingVolumeWarnings && this.boundingVolumeWarnings.length) {